requirements.txt
Django==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
pytz==2018.9
argon2-cffi==19.1.0
psycopg2-binary==2.7.7
python-dotenv==0.10.1
drfpasswordless==1.2.0
django-rest-knox==3.6.0
requests==2.21.0
faker==1.0.4
django-crispy-forms==1.7.2
django-tinymce==2.8.0
django-colorful==1.3
dj_static==0.0.6
gunicorn==19.9.0
django-dynamic-db-router==0.3.0
django-ace==1.0.5
django-ses==0.8.12
bs4==0.0.1
